refactor to make verify consumable as library
